Create editDistance.cpp
the logic here is that we have to do all the three operations and run a recursion through it and get the solution which uses the least number of operations. I have included the recursion solution also to have more clearity, but we have to apply dp for optimal solution.